fix: object horizontal filtering, vertical flip and pixel calculation
This patch properly takes into account the hardware offset applied to
the X and Y position of the object by converting the values to signed.
This allows proper horizontal filtering of the object aswell as proper
pixel calculation.

This hardware offset calculation also affects the Y flip of 8x16 objects.